backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
httpx
openai
python-dotenv
docxtpl
